+++ modules/nukkit/src/main/java/mc/alk/nukkit/NukkitServer.java
@@ -26,12 +26,12 @@ public class NukkitServer extends MCServer {
 	}
 
 	@Override
-	public MCLocation getMCLocation(String world, int x, int y, int z) {
+	public MCLocation getMCLocation(String world, double x, double y, double z) {
 		return new NukkitLocation(world, x, y, z);
 	}
 
 	@Override
-	public MCLocation getMCLocation(String world, int x, int y, int z, float pitch, float yaw) {
+	public MCLocation getMCLocation(String world, double x, double y, double z, float pitch, float yaw) {
 		return new NukkitLocation(world, x, y, z, pitch, yaw);
 	}
